# src/api/v1/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nubabel Public API
  description: |
    Public API for Nubabel - Multi-Agent Orchestration Platform.

    ## Authentication
    All endpoints require Bearer token authentication using an API key.
    Include the header: `Authorization: Bearer <your_api_key>`

    ## Rate Limiting
    Rate limits depend on your API key tier:
    - **Free**: 60 requests/minute, 1,000 requests/day
    - **Pro**: 300 requests/minute, 10,000 requests/day
    - **Enterprise**: 1,000 requests/minute, 100,000 requests/day

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit-Minute`: Your per-minute limit
    - `X-RateLimit-Remaining-Minute`: Remaining requests this minute
    - `X-RateLimit-Limit-Day`: Your per-day limit
    - `X-RateLimit-Remaining-Day`: Remaining requests today

    ## Errors
    All errors follow a consistent format:
    ```json
    {
      "error": "error_code",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Nubabel Support
    email: support@nubabel.io
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Production API
...
security:
  - BearerAuth: []
paths:
...
  /agents:
    get:
      summary: List available agents
      description: Returns all agents available in the system
      tags: [Agents]
      security:
        - BearerAuth: [agents:read]
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
  /agents/{agentId}:
    get:
      summary: Get agent details
      tags: [Agents]
      security:
        - BearerAuth: [agents:read]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AgentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /agents/{agentId}/execute:
    post:
      summary: Execute an agent
      description: Trigger agent execution with the specified input
      tags: [Agents]
      security:
        - BearerAuth: [agents:execute]
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
              properties:
                input:
                  type: string
                  description: Input prompt for the agent
                context:
                  type: object
                  description: Additional context for execution
                sessionId:
                  type: string
                  description: Session ID for tracking
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      executionId:
                        type: string
                      status:
                        type: string
                        enum: [pending, running]
                      startedAt:
                        type: string
                        format: date-time
...
  /executions:
    get:
      summary: List executions
      tags: [Executions]
      security:
        - BearerAuth: [executions:read]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: workflowId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Execution'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /executions/{executionId}:
    get:
      summary: Get execution details
      tags: [Executions]
      security:
        - BearerAuth: [executions:read]
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ExecutionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /executions/{executionId}/cancel:
    post:
      summary: Cancel execution
      tags: [Executions]
      security:
        - BearerAuth: [executions:read]
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Execution cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      status:
                        type: string
                      cancelledAt:
                        type: string
                        format: date-time
  /executions/stats/summary:
    get:
      summary: Get execution statistics
      tags: [Executions]
      security:
        - BearerAuth: [executions:read]
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
            maximum: 30
      responses:
        '200':
          description: Execution statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      period:
                        type: object
                        properties:
                          start:
                            type: string
                            format: date-time
                          end:
                            type: string
                            format: date-time
                          days:
                            type: integer
                      totals:
                        type: object
                        properties:
                          total:
                            type: integer
                          completed:
                            type: integer
                          failed:
                            type: integer
                          cancelled:
                            type: integer
                          running:
                            type: integer
                      rates:
                        type: object
                        properties:
                          successRate:
                            type: number
                          failureRate:
                            type: number
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the Developer Portal
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        version:
          type: string
    AgentDetail:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            tools:
              type: array
              items:
                type: string
            systemPrompt:
              type: string
            capabilities:
              type: array
              items:
                type: string
...
    Execution:
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        workflowName:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
    ExecutionDetail:
      allOf:
        - $ref: '#/components/schemas/Execution'
        - type: object
          properties:
            input:
              type: object
            output:
              type: object
            error:
              type: string
            nodeExecutions:
              type: array
              items:
                type: object
                properties:
                  nodeId:
                    type: string
                  status:
                    type: string
                  startedAt:
                    type: string
                    format: date-time
                  completedAt:
                    type: string
                    format: date-time
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not_found
              message:
                type: string
...